Move report api stubs in its own file
