+++ packages/rum-core/src/domain/rumEventsCollection/error/trackReportError.spec.ts
@@ -1,7 +1,8 @@
 import type { RawError, Subscription } from '@datadog/browser-core'
 import { ErrorHandling, ErrorSource, Observable, clocksNow, isChromium } from '@datadog/browser-core'
 import type { Clock } from '../../../../../core/test/specHelper'
-import { stubReportingObserver, mockClock } from '../../../../../core/test/specHelper'
+import { mockClock } from '../../../../../core/test/specHelper'
+import { stubReportingObserver } from '../../../../../core/test/stubReportApis'
 import { trackReportError } from './trackReportError'
 
 describe('trackReportError', () => {
